workflow: disable plantuml by default
The UML stuff is really only needed when updating docs, and it can be
locally enabled by the developer at that time without editing the settings
file. (Whereas if it's on by default, it can't be locally disabled without
editing the settings file.) Leaving it commented-out is the best default.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>