workflow: consolidate simple-attest
Reorganize the simple-attest definitions into a common directory,
workflow/simple-attest/, and rename some entities to better describe the
model;
   simple-attest-client -> simple-attest/client
   simple-attest-server -> simple-attest/server/ro
  simple-attest-updater -> simple-attest/server/rw
      simple-attest-git -> simple-attest/db/rw
   simple-attest-git-ro -> simple-attest/db/ro

The "db" is the authoratative database state (the "vdb" volume) with two
services/containers operating on it, one mounting it read-write (providing
the REST API to manipulate the database) and the other mounting it
read-only (providing a git-daemon for replication to attestation service
instances). These are called db.rw and db.ro to emphasize this k8s-like,
"two peas in a _pod_" quality, and we lose the "git" naming in order to
deemphasize that the underlying use of git for the DB is merely an
implementation detail.

The "server" is an instance of the attestation service, or rather its
state (the "vserver" volume), and also has two services/containers
operating on it, one read-write to replicate state from the authoritative
db, and the other read-only, which is the _actual_ attestation service!
Hence, "server-rw" and "server-ro".

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>