workflow: run swtpm in a container, and change workflow
swtpm is run in its own IMAGE, meaning the host is now a 2-container combo
combining "swtpm" and "client". In this way, it is similar to the "db" and
"server" entities which also have 2-container functions for each. (db has
a rw half for managing the DB and providing the REST API, and a ro half
for providing a replication service. server has a rw half for replicating
from the db, and a ro half that implements the attestation protocol.) So
use the mariner "service" and "group" notions to represent this and
generate rules. And use the "node" and "edge" notions to build the
dependency relationships between all the moving parts.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>