hcp: add a caboodle layer
The normal configuration for HCP is to run highly-compartmentalized and
function-specific container images for each of the HCP services and
sub-services. In the development environment, these run on a docker
network with default settings (passed in as environment variables via
docker-run) that connect them together.

The caboodle idea provides the opposite approach. A single "union"
container image that contains all the HCP functionality, and the
development environment launches this container with alternative settings
to allow the services to all be executed within the same entity and
connect to each other locally.

A test-everything.sh script is placed in the root directory of the
caboodle image that demonstrates how to start the services and run a
single invocation of the client.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>