hcp: prefix all service parameters with "HCP_"
As could be seen in the caboodle container, running "set" to scan the
environment and look for settings would show them scattered, as they were
prefixed only with the service name (ATTESTSVC_, ENROLLSVC_, etc) and
the alphabetical listing isn't friendly. This is true in the other
containers also, but less obviously annoying. This changes them all to
set, use, and expect a "HCP_" prefix, so that the parameters are easy to
find in the environment.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>