Refactor attest-server
This creates a separation between the attestation server logic (now in
attest_server_sub.py) and the details associated with the HTTP server. The
interface between them is still HTTP, but as an API call that takes the
request body and returns the response code (200, 401, ...) and body. This
should make it easier to integrate the attestation server logic with
different web-servers, frameworks, etc.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>