hcp: improve enroll.py and document a curl/jq equivalent
This script is part of swtpmsvc, because we use it to enroll a new sTPM's
EKpub as it is generated.

Importantly though, it also serves as reference code for any users/tooling
that need to hit the enrollment API programmatically. This is implemented
in python and is explanatory, but we also add (in comments at the top) a
quick-and-dirty way of performing the same operation by calling 'curl'
directly and (optionally) extracting the result code from the response
using 'jq'.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>