Convert sbin/attest-server to a flask app
This also changes the protocol (its http transport) slightly. Recall that
attestation currently consists of a single tarball input (the "quote" sent
from client to server) and a single tarball output (the "asset bundle"
sent from server back to client). Before this change, the http request
body _was_ the input tarball, which makes it hard to leverage the
request-handling infrastructure of flask (or any other web framework),
especially w.r.t. consistency checking and other defensive programming.
With this change, the http request is now a multipart "form", where a
file-valued element called "quote" is expected.

On the client side, a single argument to 'curl' is modified to support
this change, included in this commit.

On the build side, this requires an additional python module (flask) to be
present, which is added to the "requirements" target of the top-level
Makefile.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>