hcp/run: make "-it" optional for (container) functions
For services, docker-run always uses "-d". Prior to this change, for
functions, docker-run always used "-it --rm". Now, the latter only forces
the use of "--rm", so functions can add (or not) "-it" to their _ARGS
attribute if desired.

With this, "caboodle" (which runs a bash shell) runs with "-it" but
"client" (which may have stdin/stdout piped, as is the case in a CI
pipeline) does not.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>