hcp: make "clean" robust against failed Dockerfiles
We do our best to avoid having exited-but-not-removed containers lying
around. E.g. "--rm" is passed to all foregrounding docker-run commands.
However there is nothing we can do about containers started by
docker-build (when executing commands from a Dockerfile) that fail, exit,
and linger. Furthermore, this can break "make clean", because we are
unable to remove container images that are ancestors of such lingering
containers.

This adds a "preclean_hcp" rule to detect such containers and remove them.
Any "clean_*" rule that removes container images can create an order-only
("|") dependency on this rule to clear the way.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>