hcp: run enrollsvc-mgmt using 'uwsgi', and add controls
As with the recent change to make sbin/attest-server use uwsgi for running
the flask app, here we make a similar change to the "mgmt" sub-service of
enrollsvc. Environment-variable controls are declared and implemented to
allow caller-provided settings to be propagated in, following the same
conventions used in sbin/attest-server.

The other sub-service of enrollsvc ("repl", i.e. the git-daemon) is also
retrofitted to export environment-variable controls.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>