hcp/run: reorganize and improve run-time controls.
Rather than duplicating settings between the "direct" workflow (in
hcp/run/direct.mk) and the source dev/debug workflow (in hcp/Makefile), we
organize for the latter to use the former. This involves it setting
HCP_RUN_SETTINGS of course, but it also involves separating the controls
between those that identify images and paths (direct.mk will defer if they
are already set, which they will be in the source workflow) and those that
provide application configuration (direct.mk sets these unconditionally,
as you would edit it or bypass it if you intend anything else).

With this change, the full set of enrollsvc and attestsvc controls are now
available to the user of the direct workflow.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>