hcp: 'soakenroll' -> 'soak'
The resulting 'soak' command can eventually support a mix of enrollment
and attestation operations with this change. There is a --pcattest
argument that specifies, in the case of operating on an already-enrolled
EK, the probability that an attestation operation will be attempted rather
than an unenrollment. The default is zero for now, and if it is set
non-zero, the 'attestation' path is stubbed out as a no-op for now.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>