hcp: have swtpmsvc prefer unix domain sockets over TCP
For swtpm setup, a local/temporary swtpm instance is started, where it's
significantly easier and more reliable to find a private and unused path
than it is to find an unused TCP port, so use "unixio" instead.

For runtime, start the exposed swtpm instance on a unix domain socket iff
the container was given an HCP_SOCKET environment variable with the path
to use.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>