hcp: improve enrollment 'add' API
Certain orchestration flows can generate duplicate registration events for
the same host+TPM, and so it's useful to be able to distinguish this from
the error cases. To do so, we have that case return a status of 2 (other
errors return 1, and success is 0). Also, for all error cases, a single
line of output is captured and returned as a "txt" field in the JSON
returned to the caller.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>